[sbus] Design improvement  #21390
