+++ bundles/org.openhab.binding.sbus/src/main/java/org/openhab/binding/sbus/internal/handler/SbusContactHandler.java
@@ -12,8 +12,11 @@
  */
 package org.openhab.binding.sbus.internal.handler;
 
+import org.openhab.binding.sbus.BindingConstants;
 import org.openhab.binding.sbus.internal.SbusService;
+import org.openhab.binding.sbus.internal.config.ContactSensorType;
 import org.openhab.binding.sbus.internal.config.SbusChannelConfig;
+import org.openhab.binding.sbus.internal.config.SbusContactConfig;
 import org.openhab.binding.sbus.internal.config.SbusDeviceConfig;
 import org.openhab.core.library.types.OpenClosedType;
 import org.openhab.core.thing.Channel;
@@ -25,15 +28,26 @@ import org.openhab.core.types.Command;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import ro.ciprianpascu.sbus.msg.MotionSensorStatusReport;
 import ro.ciprianpascu.sbus.msg.ReadDryChannelsRequest;
 import ro.ciprianpascu.sbus.msg.ReadDryChannelsResponse;
+import ro.ciprianpascu.sbus.msg.ReadNineInOneStatusRequest;
+import ro.ciprianpascu.sbus.msg.ReadNineInOneStatusResponse;
 import ro.ciprianpascu.sbus.msg.SbusResponse;
 import ro.ciprianpascu.sbus.procimg.InputRegister;
 
 /**
- * The {@link SbusContactHandler} is responsible for handling traditional Sbus contact sensor devices.
- * It supports reading the current contact state (open/closed) using the ReadDryChannelsRequest protocol.
- * For 9-in-1 sensor devices, use the Sbus9in1ContactHandler instead.
+ * The {@link SbusContactHandler} handles Sbus contact sensor devices.
+ * It supports two protocols internally, selected via the {@code type} configuration parameter:
+ * <ul>
+ * <li>{@code 012c} (default) - traditional dry contact sensors using ReadDryChannelsRequest/Response</li>
+ * <li>{@code 02ca} - 9-in-1 multi-sensor devices using ReadNineInOneStatusRequest/Response and
+ * MotionSensorStatusReport broadcasts</li>
+ * </ul>
+ * The active protocol is (re-)resolved every time {@link #initializeChannels()} runs. Since the default
+ * {@code thingUpdated()} implementation disposes and re-initializes the same handler instance when the
+ * Thing configuration changes, changing the {@code type} parameter on an existing Thing takes effect
+ * immediately without requiring the Thing to be deleted and recreated.
  *
  * @author Ciprian Pascu - Initial contribution
  */
@@ -41,12 +55,18 @@ public class SbusContactHandler extends AbstractSbusHandler {
 
     private final Logger logger = LoggerFactory.getLogger(SbusContactHandler.class);
 
+    private volatile ContactSensorType sensorType = ContactSensorType.SENSOR_012C;
+
     public SbusContactHandler(Thing thing) {
         super(thing);
     }
 
     @Override
     protected void initializeChannels() {
+        SbusContactConfig config = getConfigAs(SbusContactConfig.class);
+        sensorType = config.getSensorType();
+        logger.debug("Initialized contact handler {} with sensor type {}", getThing().getUID(), sensorType);
+
         // Get all channel configurations from the thing
         for (Channel channel : getThing().getChannels()) {
             // Channels are already defined in thing-types.xml, just validate their configuration
@@ -59,6 +79,50 @@ public class SbusContactHandler extends AbstractSbusHandler {
 
     @Override
     protected void pollDevice() {
+        if (sensorType == ContactSensorType.MULTI_SENSOR_02CA) {
+            pollNineInOneDevice();
+        } else {
+            pollDryChannelsDevice();
+        }
+    }
+
+    @Override
+    public void handleCommand(ChannelUID channelUID, Command command) {
+        // Contact sensors are read-only
+        logger.debug("Contact device is read-only, ignoring command");
+    }
+
+    @Override
+    protected void processAsyncMessage(SbusResponse response) {
+        if (sensorType == ContactSensorType.MULTI_SENSOR_02CA) {
+            processNineInOneAsyncMessage(response);
+        } else {
+            processDryChannelsAsyncMessage(response);
+        }
+    }
+
+    @Override
+    protected boolean isMessageRelevant(SbusResponse response) {
+        SbusDeviceConfig config = getConfigAs(SbusDeviceConfig.class);
+        if (sensorType == ContactSensorType.MULTI_SENSOR_02CA) {
+            if (response instanceof MotionSensorStatusReport) {
+                // Motion sensor status reports are broadcast messages
+                return response.getSourceSubnetID() == config.subnetId && response.getSourceUnitID() == config.id;
+            } else if (response instanceof ReadNineInOneStatusResponse) {
+                return response.getSubnetID() == config.subnetId && response.getUnitID() == config.id;
+            }
+            return false;
+        } else {
+            if (response instanceof ReadDryChannelsResponse) {
+                return response.getSubnetID() == config.subnetId && response.getUnitID() == config.id;
+            }
+            return false;
+        }
+    }
+
+    // 012C Dry Contact Protocol
+
+    private void pollDryChannelsDevice() {
         final SbusService adapter = super.sbusAdapter;
         if (adapter == null) {
             updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
@@ -77,14 +141,6 @@ public class SbusContactHandler extends AbstractSbusHandler {
             logger.warn("Error polling contact device {}: {}", getThing().getUID(), e.getMessage());
         }
     }
-
-    @Override
-    public void handleCommand(ChannelUID channelUID, Command command) {
-        // Contact sensors are read-only
-        logger.debug("Contact device is read-only, ignoring command");
-    }
-
-    // SBUS Protocol Adaptation Methods
 
     /**
      * Reads contact status channel values from an SBUS device.
@@ -109,14 +165,10 @@ public class SbusContactHandler extends AbstractSbusHandler {
                     "Unexpected response type: " + (response != null ? response.getClass().getSimpleName() : "null"));
         }
 
-        boolean[] contactStates = extractContactStatuses(statusResponse);
-        return contactStates;
+        return extractContactStatuses(statusResponse);
     }
 
-    // Async Message Handling
-
-    @Override
-    protected void processAsyncMessage(SbusResponse response) {
+    private void processDryChannelsAsyncMessage(SbusResponse response) {
         try {
             if (response instanceof ReadDryChannelsResponse statusResponse) {
                 // Process status channel response using existing logic
@@ -129,16 +181,6 @@ public class SbusContactHandler extends AbstractSbusHandler {
             logger.warn("Error processing async message in contact handler {}: {}", getThing().getUID(),
                     e.getMessage());
         }
-    }
-
-    @Override
-    protected boolean isMessageRelevant(SbusResponse response) {
-        if (response instanceof ReadDryChannelsResponse) {
-            // Traditional contact sensor messages
-            SbusDeviceConfig config = getConfigAs(SbusDeviceConfig.class);
-            return response.getSubnetID() == config.subnetId && response.getUnitID() == config.id;
-        }
-        return false;
     }
 
     /**
@@ -171,5 +213,131 @@ public class SbusContactHandler extends AbstractSbusHandler {
             statuses[i] = (registers[i].getValue() & 0xff) > 0; // Convert to boolean
         }
         return statuses;
+    }
+
+    // 02CA / 9-in-1 Multi-Sensor Protocol
+
+    private void pollNineInOneDevice() {
+        final SbusService adapter = super.sbusAdapter;
+        if (adapter == null) {
+            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.CONFIGURATION_ERROR,
+                    "@text/error.device.adapter-not-initialized");
+            return;
+        }
+
+        try {
+            SbusDeviceConfig config = getConfigAs(SbusDeviceConfig.class);
+            ReadNineInOneStatusResponse response = readNineInOneStatus(adapter, config.subnetId, config.id);
+
+            // Update all contact channels from the response
+            updateContactChannelsFromResponse(response);
+
+            updateStatus(ThingStatus.ONLINE);
+        } catch (IllegalStateException e) {
+            updateStatus(ThingStatus.OFFLINE, ThingStatusDetail.COMMUNICATION_ERROR,
+                    "@text/error.device.communication");
+            logger.warn("Error polling 9-in-1 contact sensor {}: {}", getThing().getUID(), e.getMessage());
+        }
+    }
+
+    /**
+     * Read 9-in-1 sensor status from device.
+     *
+     * @param adapter the SBUS service adapter
+     * @param subnetId the subnet ID
+     * @param deviceId the device ID
+     * @return ReadNineInOneStatusResponse
+     * @throws IllegalStateException if communication fails
+     */
+    private ReadNineInOneStatusResponse readNineInOneStatus(SbusService adapter, int subnetId, int deviceId)
+            throws IllegalStateException {
+        ReadNineInOneStatusRequest request = new ReadNineInOneStatusRequest();
+        request.setSubnetID(subnetId);
+        request.setUnitID(deviceId);
+
+        SbusResponse response = adapter.executeTransaction(request);
+        if (!(response instanceof ReadNineInOneStatusResponse statusResponse)) {
+            throw new IllegalStateException(
+                    "Unexpected response type: " + (response != null ? response.getClass().getSimpleName() : "null"));
+        }
+
+        return statusResponse;
+    }
+
+    private void processNineInOneAsyncMessage(SbusResponse response) {
+        try {
+            if (response instanceof MotionSensorStatusReport report) {
+                // Process motion sensor status report for dry contact updates
+                updateContactChannelsFromReport(report);
+                updateStatus(ThingStatus.ONLINE);
+                logger.debug("Processed async motion sensor status report for 9-in-1 contact handler {}",
+                        getThing().getUID());
+            }
+        } catch (IllegalStateException | IllegalArgumentException e) {
+            logger.warn("Error processing async message in 9-in-1 contact handler {}: {}", getThing().getUID(),
+                    e.getMessage());
+        }
+    }
+
+    /**
+     * Update contact channels from ReadNineInOneStatusResponse.
+     *
+     * @param response the 9-in-1 status response
+     */
+    private void updateContactChannelsFromResponse(ReadNineInOneStatusResponse response) {
+        for (Channel channel : getThing().getChannels()) {
+            ChannelUID channelUID = channel.getUID();
+            var channelTypeUID = channel.getChannelTypeUID();
+            if (channelTypeUID != null && BindingConstants.CHANNEL_TYPE_CONTACT.equals(channelTypeUID.getId())) {
+                SbusChannelConfig channelConfig = channel.getConfiguration().as(SbusChannelConfig.class);
+
+                // Use channelNumber to determine which dry contact (1 or 2, default to 1)
+                int channelNumber = channelConfig.channelNumber > 0 ? channelConfig.channelNumber : 1;
+                boolean contactState = false;
+
+                if (channelNumber == 1) {
+                    contactState = response.getDryContact1Status() > 0;
+                } else if (channelNumber == 2) {
+                    contactState = response.getDryContact2Status() > 0;
+                }
+
+                OpenClosedType state = contactState ? OpenClosedType.OPEN : OpenClosedType.CLOSED;
+                updateState(channelUID, state);
+
+                logger.debug("Updated 9-in-1 contact channel {} (number {}) state: {}", channelUID.getId(),
+                        channelNumber, state);
+            }
+        }
+    }
+
+    /**
+     * Update contact channels from MotionSensorStatusReport.
+     *
+     * @param report the motion sensor status report
+     */
+    private void updateContactChannelsFromReport(MotionSensorStatusReport report) {
+        for (Channel channel : getThing().getChannels()) {
+            ChannelUID channelUID = channel.getUID();
+            var channelTypeUID = channel.getChannelTypeUID();
+            if (channelTypeUID != null && BindingConstants.CHANNEL_TYPE_CONTACT.equals(channelTypeUID.getId())) {
+                SbusChannelConfig channelConfig = channel.getConfiguration().as(SbusChannelConfig.class);
+
+                // Use channelNumber to determine which dry contact (1 or 2, default to 1)
+                int channelNumber = channelConfig.channelNumber > 0 ? channelConfig.channelNumber : 1;
+                boolean contactState = false;
+
+                if (channelNumber == 1) {
+                    contactState = report.getDryContactStatus(0) > 0; // First dry contact (index 0)
+                } else if (channelNumber == 2) {
+                    contactState = report.getDryContactStatus(1) > 0; // Second dry contact (index 1)
+                }
+
+                OpenClosedType state = contactState ? OpenClosedType.OPEN : OpenClosedType.CLOSED;
+                updateState(channelUID, state);
+
+                logger.debug("Updated 9-in-1 contact channel {} (number {}) state from report: {}", channelUID.getId(),
+                        channelNumber, state);
+            }
+        }
     }
 }
