+++ bundles/org.openhab.binding.sbus/src/main/java/org/openhab/binding/sbus/internal/handler/SbusHandlerFactory.java
@@ -18,8 +18,6 @@ import java.util.Set;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.binding.sbus.internal.config.ContactSensorType;
-import org.openhab.binding.sbus.internal.config.SbusContactConfig;
 import org.openhab.core.thing.Bridge;
 import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingTypeUID;
@@ -69,17 +67,8 @@ public class SbusHandlerFactory extends BaseThingHandlerFactory {
             logger.debug("Creating Sbus RGBW handler for thing {}", thing.getUID());
             return new SbusRgbwHandler(thing);
         } else if (thingTypeUID.equals(THING_TYPE_CONTACT_SENSOR)) {
-            // Determine which contact handler to create based on sensor type configuration
-            SbusContactConfig config = thing.getConfiguration().as(SbusContactConfig.class);
-            ContactSensorType sensorType = config.getSensorType();
-
-            if (sensorType == ContactSensorType.MULTI_SENSOR_02CA) {
-                logger.debug("Creating Sbus 02CA multi-sensor contact handler for thing {}", thing.getUID());
-                return new Sbus9in1ContactHandler(thing);
-            } else {
-                logger.debug("Creating Sbus 012C contact sensor handler for thing {}", thing.getUID());
-                return new SbusContactHandler(thing);
-            }
+            logger.debug("Creating Sbus contact sensor handler for thing {}", thing.getUID());
+            return new SbusContactHandler(thing);
         } else if (thingTypeUID.equals(THING_TYPE_MOTION_SENSOR)) {
             logger.debug("Creating Sbus motion sensor handler for thing {}", thing.getUID());
             return new SbusMotionSensorHandler(thing);
